+++ CHANGES.md
@@ -19,4 +19,5 @@ Changelog
 * Improved logging.
 * Improved documentation.
 * Easier unregistration/shutdown using either `java.lang.Runtime` shutdown hooks or customized shutdown systems (Issue #20).
-* Utilities for shutdown in a servlet environment. 
+* Utilities for shutdown in a servlet environment (Issue #21). 
+* ImmutableProperties payload object (Issue #40).